// app/models/planetsResultApiData.ts
export type PlanetsResultApiData = {
  name: string;
  rotation_period: string;
  orbital_period: string;
  diameter: string;
  climate: string;
  gravity: string;
  terrain: string;
  surface_water: string;
  population: string;
  residents: any;
  films: any;
  created: Date;
  edited: Date;
  url: string;
};